Core/AuctionHouse: After taking a break at the start of patch 8.3 auction house is now back in business.
Port From (https://github.com/TrinityCore/TrinityCore/commit/aced88b09bd918b0ed17b6c5e8a6048788600d9d)

+++ Source/Game/Network/Packets/MailPackets.cs
@@ -326,7 +326,7 @@ namespace Game.Network.Packets
                 if (item.GetEnchantmentId(slot) == 0)
                     continue;
 
-                Enchants.Add(new ItemEnchantData((int)item.GetEnchantmentId(slot), item.GetEnchantmentDuration(slot), (int)item.GetEnchantmentCharges(slot), (byte)slot));
+                Enchants.Add(new ItemEnchantData(item.GetEnchantmentId(slot), item.GetEnchantmentDuration(slot), (int)item.GetEnchantmentCharges(slot), (byte)slot));
             }
 
             byte i = 0;
